Agregado detalles de pelicula seleccionada dinámicos y cambiado modal de archivo
